Create Profits Index View
We need a root page for our application, so it's necessary the
profits controller with the index action and its view. The first step
is call the profits index in our router like the root page by default
then an application controller was added to manage the common methods
for all controllers and finally the profits controller was added
rendering the index view.